Наводит порядок в функциях
